requirements.txt
google-genai
google-api-python-client
google-auth
logfire
pydantic
